create PublicationCreation.page.js and add selectors
